# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5f512d2467541c707d17b59d
  title: TESTAPP
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  /company:
    post:
      x-skaffolder-id: 5f5139f667541c707d17b68f
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5f5132be67541c707d17b657
      x-skaffolder-resource: CompanyInfo
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - CompanyInfo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyInfo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Description:
                  type: string
                Facebook:
                  type: string
                HQAddress:
                  type: string
                LinkedIn:
                  type: string
                Name:
                  type: string
                Twitter:
                  type: string
                VATID:
                  type: string
                Website:
                  type: string
    get:
      x-skaffolder-id: 5f5139f667541c707d17b692
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5f5132be67541c707d17b657
      x-skaffolder-resource: CompanyInfo
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - CompanyInfo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyInfo"
  "/company/{id}":
    delete:
      x-skaffolder-id: 5f5139f667541c707d17b691
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f5132be67541c707d17b657
      x-skaffolder-resource: CompanyInfo
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - CompanyInfo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f5139f667541c707d17b68b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f5132be67541c707d17b657
      x-skaffolder-resource: CompanyInfo
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - CompanyInfo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyInfo"
    post:
      x-skaffolder-id: 5f5139f667541c707d17b68c
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f5132be67541c707d17b657
      x-skaffolder-resource: CompanyInfo
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - CompanyInfo
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyInfo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInfo"
...
  /user:
    post:
      x-skaffolder-id: 5f512d2467541c707d17b5ae
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5f512d2467541c707d17b5a7
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Email:
                  type: string
                ProfileType:
                  type: string
                RegistrationStatus:
                  type: boolean
                RoleType:
                  type: string
                SubscriptionType:
                  type: string
                UserType:
                  type: string
                password:
                  type: string
    get:
      x-skaffolder-id: 5f512d2467541c707d17b5b0
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5f512d2467541c707d17b5a7
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5f512d2467541c707d17b5ab
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f512d2467541c707d17b5a7
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f512d2467541c707d17b5ad
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f512d2467541c707d17b5a7
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5f512d2467541c707d17b5ac
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f512d2467541c707d17b5a7
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CompanyInfo:
      x-skaffolder-id: 5f5132be67541c707d17b657
      x-skaffolder-id-db: 5f512d2467541c707d17b59e
      x-skaffolder-id-entity: 5f5132be67541c707d17b64e
      x-skaffolder-url: /company
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Description:
          type: string
          x-skaffolder-id-attr: 5f5132be67541c707d17b650
          x-skaffolder-type: String
        Facebook:
          type: string
          x-skaffolder-id-attr: 5f5132be67541c707d17b656
          x-skaffolder-type: String
        HQAddress:
          type: string
          x-skaffolder-id-attr: 5f5132be67541c707d17b652
          x-skaffolder-type: String
        LinkedIn:
          type: string
          x-skaffolder-id-attr: 5f5132be67541c707d17b654
          x-skaffolder-type: String
        Name:
          type: string
          x-skaffolder-id-attr: 5f5132be67541c707d17b64f
          x-skaffolder-type: String
        Twitter:
          type: string
          x-skaffolder-id-attr: 5f5132be67541c707d17b655
          x-skaffolder-type: String
        VATID:
          type: string
          x-skaffolder-id-attr: 5f5132be67541c707d17b651
          x-skaffolder-type: String
        Website:
          type: string
          x-skaffolder-id-attr: 5f5132be67541c707d17b653
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
    User:
      x-skaffolder-id: 5f512d2467541c707d17b5a7
      x-skaffolder-id-db: 5f512d2467541c707d17b59e
      x-skaffolder-id-entity: 5f512d2467541c707d17b59f
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Email:
          type: string
          x-skaffolder-id-attr: 5f512d2467541c707d17b5a2
          x-skaffolder-type: String
        ProfileType:
          type: string
          x-skaffolder-id-attr: 5f51364767541c707d17b66c
          x-skaffolder-type: String
        RegistrationStatus:
          type: boolean
          x-skaffolder-id-attr: 5f51341267541c707d17b65d
          x-skaffolder-type: Boolean
        RoleType:
          type: string
          x-skaffolder-id-attr: 5f51364767541c707d17b66b
          x-skaffolder-type: String
        SubscriptionType:
          type: string
          x-skaffolder-id-attr: 5f51368467541c707d17b66d
          x-skaffolder-type: String
        UserType:
          type: string
          x-skaffolder-id-attr: 5f5134a867541c707d17b65e
          x-skaffolder-type: String
          x-skaffolder-required: true
        password:
          type: string
          x-skaffolder-id-attr: 5f512d2467541c707d17b5a1
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - UserType
        - password
...