# src/main/java/com/testapp/db/testapp_db/dtos/CompanyInfoDto.java
package com.testapp.db.testapp_db.dtos;

import com.testapp.db.testapp_db.dtos.base.CompanyInfoBaseDto;

public class CompanyInfoDto extends CompanyInfoBaseDto {

	//OVERRIDE HERE YOUR CUSTOMIZATIONS
}